Refactor code to add new books, authors and locations
This commit refactors the code to add new books, authors and locations. It also updates the author-book relationship by adding a book to an author's set of books. The location object is now created with a null book parameter before being saved in the database. Finally, some commented out lines of code have been removed from the BookController class.